Add support for linking stable names (#51)
Intended to be used as `[stable.name]{.sref}`, and transforms that into
`1.2.3.4 [stable.name]` with [...] linking to 'https://wg21.link/stable.name'.

`sref` comes from the XML tag used for this in LWG's issue list software.

To support the subclause number lookup, this downloads a so-called "Annex F"
file containing mapping between the stable name and the section number.
The name `annex-f` and the incantation to generate it comes from the LWG issues
documentation. https://github.com/timsong-cpp/cppwp/annex-f hosts this file.