Add an "old" attribute to the pnum class to indicate paragraph number changes
This does not effect the markup, and allows a purely css highlight model for
the changes. We intentionally use the ::before pseudo element to inject the
content so that it does not screw up selection of text (e.g. only the "real"
text of the change is eligible for selection).